cpp: merge test_array.cpp into test_datapipeline.cc
