+++ bindings/cpp/CMakeLists.txt
@@ -61,11 +61,11 @@ add_executable(test_datapipeline
   test_datapipeline.cc)
 target_link_libraries(test_datapipeline data_pipeline_cpp)
 
-#
-add_executable(test_array  test_array.cpp)
-target_link_libraries(test_array PUBLIC pybind11::embed)
+#  array test code has moved into test_datapipeline.cc
+#add_executable(test_array  test_array.cpp)
+#target_link_libraries(test_array PUBLIC pybind11::embed)
 
-foreach(MyTarget test_array data_pipeline_cpp test_datapipeline)
+foreach(MyTarget data_pipeline_cpp test_datapipeline)
     # Let's set compiler-specific flags
     if (${CMAKE_CXX_COMPILER_ID} STREQUAL "GNU")
         target_compile_options(${MyTarget} PRIVATE -fvisibility=hidden)
